+++ src/test/groovy/edu/ie3/datamodel/io/factory/FactoryTest.groovy
@@ -6,6 +6,7 @@
 package edu.ie3.datamodel.io.factory
 
 import edu.ie3.datamodel.exceptions.FactoryException
+import edu.ie3.datamodel.models.input.EmInput
 import edu.ie3.datamodel.utils.Try
 import spock.lang.Shared
 import spock.lang.Specification
@@ -17,7 +18,7 @@ class FactoryTest extends Specification {
 
   def "A Factory can return unused fields correctly"() {
     when:
-    def unused = factory.getUnusedFields(actualFields as Set<String>, validFieldSets)
+    def unused = factory.getUnusedFields(actualFields as Set<String>, validFieldSets, [] as Set<String>)
 
     then:
     unused == expected as Set<String>
@@ -96,6 +97,15 @@ class FactoryTest extends Specification {
         ["uuid", "value1", "value3"] as Set<String>,
         ["id", "time", "value1"] as Set<String>
       ]
+    }
+
+    @Override
+    protected Map<String, String> getReplacedFields(Class<?> entityClass) {
+      if (entityClass.isAssignableFrom(EmInput)) {
+        return ["parentEm": "controllingEM"]
+      } else {
+        return [:]
+      }
     }
   }
 }